refactor: extract reconcile helpers into dedicated files
Move pure helper functions out of the reconcile files and into
domain-specific helper files:

- reconcile_helpers.go: shared helpers (ReconcileSnapshot, now, isReleased)
- movie_reconcile_helpers.go: initialMovieState, findMatchingMovieResult
- series_reconcile_helpers.go: initialSeriesState, findMatchingSeriesResult,
  getSeasonRuntime, determineSeasonState, determineSeriesState

Move corresponding test functions into matching _helpers_test.go files.
